Add heaps of new types, and autodetect.
Heaps of new types have been added to select from. (Too many to list in this small section)

The name can now be autodetected.
if you input a file like `.rs`, it will automatically select `Rust`.
If you wish to override this, select a value from the dropdown.
To undo an override, press the chain button next ot the dropdown.