Auto type selection.
the file type is now somewhat inferred.

Originally to set a separate type from the extension you would do the following:

1. change the extension
2. change the name

and then resetting, you would press the paper-clip icon, to match the type in the name.

Now, it simply detects if you modify the name, or the extension:

here is an example:

`main.py`

`main` is the name
`py` is the extension

Changing the name, will not update the extension/type.

Changing the extension, will update the type.